apps/api/src/crawler/requirements.txt
fastapi==0.100.0
uvicorn==0.20.0
playwright==1.35.0
apscheduler==3.10.1
python-dotenv==1.0.0
pydantic==1.10.12